pyproject.toml
[project]
name = "medical-audit"
version = "1.0.0"
description = "Healthcare document audit system for Colombian hospitals"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client==2.190.0",
    "google-auth-oauthlib==1.2.4",
    "ocrmypdf>=17.3.0",
    "openpyxl==3.1.5",
    "pandas==2.2.2",
    "playwright>=1.58.0",
    "PyMuPDF==1.24.9",
    "python-dotenv==1.2.1",
    "streamlit>=1.40.0",
    "tqdm>=4.67.3",
]
